chainlit_bot/pyproject.toml
[project]
name = "chainlit-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chainlit>=2.5.5",
    "google-generativeai>=0.8.5",
    "openai-agents>=0.0.17",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "rich>=14.0.0",
]
